moved searchbar inside of search books page
