+++ src/components/views/Booktrak/BooktrakHome.tsx
@@ -14,6 +14,7 @@ const BooktrakHome = () => {
   const wso = useAppSelector(getWSO);
   const navigateTo = useNavigate();
   const [searchParams] = useSearchParams();
+  const [query, updateQuery] = useState("");
 
   const [results, updateResults] = useState<ModelsBook[]>([]);
   const perPage = 20;
@@ -23,6 +24,14 @@ const BooktrakHome = () => {
   const [selectedBook, updatedSelectedBook] = useState<ModelsBook | undefined>(
     undefined
   );
+
+  useEffect(() => {
+    if (searchParams?.get("q")) {
+      updateQuery(searchParams.get("q") ?? "");
+    } else {
+      updateQuery("");
+    }
+  }, [searchParams]);
 
   const loadBooks = async () => {
     if (!searchParams?.get("q")) {
@@ -186,25 +195,44 @@ const BooktrakHome = () => {
     return ListView();
   };
 
-  if (!searchParams?.get("q")) {
-    return null;
-  }
-  // This will act as a loading buffer
-  if (!results) {
-    return (
-      <article className="facebook-results">
-        <section>
-          <br />
-          <h1 className="no-matches-found">Loading...</h1>
-        </section>
-      </article>
-    );
-  }
+  const submitHandler: React.FormEventHandler<HTMLFormElement> = (event) => {
+    event.preventDefault();
+    searchParams.set("q", query);
+    navigateTo(`/booktrak?${searchParams.toString()}`);
+  };
 
   return (
-    <article className="facebook-results">
-      <section>{BooktrakResults()}</section>
-    </article>
+    <>
+      <form onSubmit={submitHandler}>
+        <input
+          id="search"
+          type="search"
+          placeholder="Search Booktrak"
+          autoFocus
+          onChange={(event) => updateQuery(event.target.value)}
+          value={query}
+        />
+        <input
+          data-disable-with="Search"
+          type="submit"
+          value="Search"
+          className="submit"
+        />
+      </form>
+      {searchParams?.get("q") &&
+        (results ? (
+          <article className="facebook-results">
+            <section>{BooktrakResults()}</section>
+          </article>
+        ) : (
+          <article className="facebook-results">
+            <section>
+              <br />
+              <h1 className="no-matches-found">Loading...</h1>
+            </section>
+          </article>
+        ))}
+    </>
   );
 };
