+++ src/components/views/Booktrak/BooktrakLayout.tsx
@@ -1,33 +1,15 @@
 // React imports
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 // Redux imports
 import { useAppSelector } from "../../../lib/store";
 import { getCurrUser } from "../../../lib/authSlice";
 
 // Additional imports
-import { Link, useSearchParams, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 const BooktrakLayout = ({ children }: { children: React.ReactElement }) => {
   const currUser = useAppSelector(getCurrUser);
-  const navigateTo = useNavigate();
-  const [searchParams] = useSearchParams();
-  const [query, updateQuery] = useState("");
-
-  useEffect(() => {
-    if (searchParams?.get("q")) {
-      updateQuery(searchParams.get("q") ?? "");
-    } else {
-      updateQuery("");
-    }
-  }, [searchParams]);
-
-  // Handles submissions
-  const submitHandler: React.FormEventHandler<HTMLFormElement> = (event) => {
-    event.preventDefault();
-    searchParams.set("q", query);
-    navigateTo(`/booktrak?${searchParams.toString()}`);
-  };
 
   return (
     <div className="facebook">
@@ -58,22 +40,6 @@ const BooktrakLayout = ({ children }: { children: React.ReactElement }) => {
                 ]}
           </ul>
         </div>
-        <form onSubmit={submitHandler}>
-          <input
-            id="search"
-            type="search"
-            placeholder="Search Booktrak"
-            autoFocus
-            onChange={(event) => updateQuery(event.target.value)}
-            value={query}
-          />
-          <input
-            data-disable-with="Search"
-            type="submit"
-            value="Search"
-            className="submit"
-          />
-        </form>
       </header>
       {children}
     </div>
